[v1.06][RISCV] Add latest extensions and refactor the mask structure
There are *a lot* of new extensions in the Linux kernel now (Linux
v6.16) since the last time I checked (Linux v6.2.10), so added them all
plus ime, which is not present in the kernel apparently, but that I
spotted in a RISC-V chip.

Previously, the mask was represented with a uint64_t value, which is
fine as long as we have less than 65 possible extensions. This is not
the case anymore, so instead I replaced that with a pointer of bool,
which gets allocated depending on the max number of extensions that
we support.